# Conflicts:
#	src/main/java/io/flowset/control/action/TestEngineConnectionAction.java
#	src/main/java/io/flowset/control/entity/incident/HistoricIncidentData.java
#	src/main/java/io/flowset/control/entity/incident/IncidentData.java
#	src/main/java/io/flowset/control/service/engine/impl/EngineUiServiceImpl.java
#	src/main/java/io/flowset/control/view/bpmengine/BpmEngineDetailView.java

+++ src/main/java/io/flowset/control/entity/ExternalTaskData.java
@@ -48,7 +48,7 @@ public class ExternalTaskData {
 
     protected String topicName;
 
-    protected Integer priority;
+    protected Long priority;
 
     protected String businessKey;
 
@@ -68,11 +68,11 @@ public class ExternalTaskData {
         this.businessKey = businessKey;
     }
 
-    public Integer getPriority() {
+    public Long getPriority() {
         return priority;
     }
 
-    public void setPriority(Integer priority) {
+    public void setPriority(Long priority) {
         this.priority = priority;
     }
